Fixed restoration of stdout-stderr in common.restore_stdout
Summary:
sys.stdout, sys.stderr may not be equal to
```
sys.__stdout__
```
,

```
sys.__stderr__
```
when the function common.redirect_stdout() is called,
especially in Python IDEs such as spyder. As a result one needs to record
the pointer to sys.stdout, sys.stderr when redirect_stdout is called,
to restore them when restore_stdout is called.

Test Plan:
Trained a tensorflow model with a CSVLogger callback, in three ways:
1- Running "python" in a console with a script filename argument
2- Running that same training script within Spyder (this part fails with the CSVLogger)
3- Performing the training in the ML plugin with the bokeh GUI and file logging in the survey
folder.

Reviewers: wayne

Reviewed By: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14878